pyproject.toml
[project]
name = "financeanalysispy"
version = "0.1.0"
description = "Financial analysis system with technical and fundamental analysis services"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "arch>=7.2.0",
    "asttokens==3.0.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "colorama==0.4.6",
    "comm==0.2.2",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "cython>=3.0.12",
    "dateparser==1.2.1",
    "decorator==5.2.1",
    "dill==0.3.9",
    "executing==2.2.0",
    "fastapi>=0.115.12",
    "financepy==0.370",
    "fonttools==4.57.0",
    "idna==3.10",
    "imageio==2.37.0",
    "ipython==9.1.0",
    "ipython-pygments-lexers==1.1.1",
    "ipywidgets==8.1.5",
    "jedi==0.19.2",
    "joblib==1.4.2",
    "jupyterlab-widgets==3.0.13",
    "kaleido>=0.2.1",
    "kiwisolver==1.4.8",
    "lean",
    "llvmlite==0.44.0",
    "matplotlib==3.10.1",
    "matplotlib-inline==0.1.7",
    "mlflow>=2.22.0",
    "mplfinance>=0.12.10b0",
    "mypy-extensions==1.0.0",
    "numba==0.61.0",
    "numpy==2.1.3",
    "packaging==24.2",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pillow==11.1.0",
    "plotly==5.24.1",
    "prettytable==3.16.0",
    "prompt-toolkit==3.0.50",
    "prophet>=1.1.6",
    "pydantic>=2.0",
    "pure-eval==0.2.3",
    "pyarrow>=19.0.1",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "pytest>=8.3.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.1.0",
    "pytz==2025.2",
    "pyyaml>=6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "schedule==1.2.2",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "seaborn==0.13.2",
    "setuptools==78.1.0",
    "six==1.17.0",
    "stack-data==0.6.3",
    "statsmodels>=0.14.4",
    "streamlit>=1.45.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "torch>=2.7.0",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "urllib3==2.3.0",
    "uvicorn[standard]>=0.34.0",
    "vectorbt==0.27.2",
    "wcwidth==0.2.13",
    "widgetsnbextension==4.0.13",
    "yfinance>=0.2.55",
]

[project.optional-dependencies]
# Optional TA-Lib support for advanced technical indicators
# Note: Requires platform-specific wheels or system installation
talib = [
    "ta-lib>=0.4.0",
]

[tool.uv.sources]
# Platform-specific ta-lib wheels (optional dependency)
# To install: uv sync --extra talib
# Windows: uses pre-built wheel in lib/
# Linux: requires system ta-lib C library installation first
#        See: https://github.com/TA-Lib/ta-lib-python#installation
ta-lib = { path = "lib/ta_lib-0.6.3-cp313-cp313-win_amd64.whl", marker = "sys_platform == 'win32'" }
